Python code for temperature an humidity
In this code, i am using the Adafruit_DHT library to communicate with the DHT11 sensor connected to pin 4 of the Raspberry Pi. We are then reading the temperature and humidity values from the sensor and checking if they exceed the threshold values. If the temperature and humidity are both above their respective thresholds, an alarm message is printed to the console. You can modify the code to include your own alarm mechanism, such as sending an email or SMS notification. The code then waits for 5 seconds before reading the sensor again.